# WIP: Optimize seterr