--- a/gplearn/functions.py
+++ b/gplearn/functions.py
@@ -123,8 +123,7 @@ def make_function(function, name, arity, wrap=True):
 
 def _protected_division(x1, x2):
     """Closure of division (x1/x2) for zero denominator."""
-    with np.errstate(divide='ignore', invalid='ignore'):
-        return np.where(np.abs(x2) > 0.001, np.divide(x1, x2), 1.)
+    return np.where(np.abs(x2) > 0.001, np.divide(x1, x2), 1.)
 
 
 def _protected_sqrt(x1):
@@ -134,20 +133,17 @@ def _protected_sqrt(x1):
 
 def _protected_log(x1):
     """Closure of log for zero arguments."""
-    with np.errstate(divide='ignore', invalid='ignore'):
-        return np.where(np.abs(x1) > 0.001, np.log(np.abs(x1)), 0.)
+    return np.where(np.abs(x1) > 0.001, np.log(np.abs(x1)), 0.)
 
 
 def _protected_inverse(x1):
     """Closure of log for zero arguments."""
-    with np.errstate(divide='ignore', invalid='ignore'):
-        return np.where(np.abs(x1) > 0.001, 1. / x1, 0.)
+    return np.where(np.abs(x1) > 0.001, 1. / x1, 0.)
 
 
 def _sigmoid(x1):
     """Special case of logistic function to transform to probabilities."""
-    with np.errstate(over='ignore', under='ignore'):
-        return 1 / (1 + np.exp(-x1))
+    return 1 / (1 + np.exp(-x1))
 
 
 add2 = _Function(function=np.add, name='add', arity=2)