--- a/gplearn/genetic.py
+++ b/gplearn/genetic.py
@@ -68,88 +68,89 @@ def _parallel_evolve(n_programs, parents, X, y, sample_weight, seeds, params):
     # Build programs
     programs = []
 
-    for i in range(n_programs):
-
-        random_state = check_random_state(seeds[i])
-
-        if parents is None:
-            program = None
-            genome = None
-        else:
-            method = random_state.uniform()
-            parent, parent_index = _tournament()
-
-            if method < method_probs[0]:
-                # crossover
-                donor, donor_index = _tournament()
-                program, removed, remains = parent.crossover(donor.program,
-                                                             random_state)
-                genome = {'method': 'Crossover',
-                          'parent_idx': parent_index,
-                          'parent_nodes': removed,
-                          'donor_idx': donor_index,
-                          'donor_nodes': remains}
-            elif method < method_probs[1]:
-                # subtree_mutation
-                program, removed, _ = parent.subtree_mutation(random_state)
-                genome = {'method': 'Subtree Mutation',
-                          'parent_idx': parent_index,
-                          'parent_nodes': removed}
-            elif method < method_probs[2]:
-                # hoist_mutation
-                program, removed = parent.hoist_mutation(random_state)
-                genome = {'method': 'Hoist Mutation',
-                          'parent_idx': parent_index,
-                          'parent_nodes': removed}
-            elif method < method_probs[3]:
-                # point_mutation
-                program, mutated = parent.point_mutation(random_state)
-                genome = {'method': 'Point Mutation',
-                          'parent_idx': parent_index,
-                          'parent_nodes': mutated}
+    with np.errstate(all='ignore'):
+        for i in range(n_programs):
+    
+            random_state = check_random_state(seeds[i])
+    
+            if parents is None:
+                program = None
+                genome = None
             else:
-                # reproduction
-                program = parent.reproduce()
-                genome = {'method': 'Reproduction',
-                          'parent_idx': parent_index,
-                          'parent_nodes': []}
-
-        program = _Program(function_set=function_set,
-                           arities=arities,
-                           init_depth=init_depth,
-                           init_method=init_method,
-                           n_features=n_features,
-                           metric=metric,
-                           transformer=transformer,
-                           const_range=const_range,
-                           p_point_replace=p_point_replace,
-                           parsimony_coefficient=parsimony_coefficient,
-                           feature_names=feature_names,
-                           random_state=random_state,
-                           program=program)
-
-        program.parents = genome
-
-        # Draw samples, using sample weights, and then fit
-        if sample_weight is None:
-            curr_sample_weight = np.ones((n_samples,))
-        else:
-            curr_sample_weight = sample_weight.copy()
-        oob_sample_weight = curr_sample_weight.copy()
-
-        indices, not_indices = program.get_all_indices(n_samples,
-                                                       max_samples,
-                                                       random_state)
-
-        curr_sample_weight[not_indices] = 0
-        oob_sample_weight[indices] = 0
-
-        program.raw_fitness_ = program.raw_fitness(X, y, curr_sample_weight)
-        if max_samples < n_samples:
-            # Calculate OOB fitness
-            program.oob_fitness_ = program.raw_fitness(X, y, oob_sample_weight)
-
-        programs.append(program)
+                method = random_state.uniform()
+                parent, parent_index = _tournament()
+    
+                if method < method_probs[0]:
+                    # crossover
+                    donor, donor_index = _tournament()
+                    program, removed, remains = parent.crossover(donor.program,
+                                                                 random_state)
+                    genome = {'method': 'Crossover',
+                              'parent_idx': parent_index,
+                              'parent_nodes': removed,
+                              'donor_idx': donor_index,
+                              'donor_nodes': remains}
+                elif method < method_probs[1]:
+                    # subtree_mutation
+                    program, removed, _ = parent.subtree_mutation(random_state)
+                    genome = {'method': 'Subtree Mutation',
+                              'parent_idx': parent_index,
+                              'parent_nodes': removed}
+                elif method < method_probs[2]:
+                    # hoist_mutation
+                    program, removed = parent.hoist_mutation(random_state)
+                    genome = {'method': 'Hoist Mutation',
+                              'parent_idx': parent_index,
+                              'parent_nodes': removed}
+                elif method < method_probs[3]:
+                    # point_mutation
+                    program, mutated = parent.point_mutation(random_state)
+                    genome = {'method': 'Point Mutation',
+                              'parent_idx': parent_index,
+                              'parent_nodes': mutated}
+                else:
+                    # reproduction
+                    program = parent.reproduce()
+                    genome = {'method': 'Reproduction',
+                              'parent_idx': parent_index,
+                              'parent_nodes': []}
+    
+            program = _Program(function_set=function_set,
+                               arities=arities,
+                               init_depth=init_depth,
+                               init_method=init_method,
+                               n_features=n_features,
+                               metric=metric,
+                               transformer=transformer,
+                               const_range=const_range,
+                               p_point_replace=p_point_replace,
+                               parsimony_coefficient=parsimony_coefficient,
+                               feature_names=feature_names,
+                               random_state=random_state,
+                               program=program)
+    
+            program.parents = genome
+    
+            # Draw samples, using sample weights, and then fit
+            if sample_weight is None:
+                curr_sample_weight = np.ones((n_samples,))
+            else:
+                curr_sample_weight = sample_weight.copy()
+            oob_sample_weight = curr_sample_weight.copy()
+    
+            indices, not_indices = program.get_all_indices(n_samples,
+                                                           max_samples,
+                                                           random_state)
+    
+            curr_sample_weight[not_indices] = 0
+            oob_sample_weight[indices] = 0
+    
+            program.raw_fitness_ = program.raw_fitness(X, y, curr_sample_weight)
+            if max_samples < n_samples:
+                # Calculate OOB fitness
+                program.oob_fitness_ = program.raw_fitness(X, y, oob_sample_weight)
+    
+            programs.append(program)
 
     return programs
 
@@ -560,8 +561,8 @@ class BaseSymbolic(BaseEstimator, metaclass=ABCMeta):
             if self.metric == 'spearman':
                 evaluation = np.apply_along_axis(rankdata, 1, evaluation)
 
-            with np.errstate(divide='ignore', invalid='ignore'):
-                correlations = np.abs(np.corrcoef(evaluation))
+
+            correlations = np.abs(np.corrcoef(evaluation))
             np.fill_diagonal(correlations, 0.)
             components = list(range(self.hall_of_fame))
             indices = list(range(self.hall_of_fame))