--- a/setup.py
+++ b/setup.py
@@ -35,4 +35,6 @@ setup(name='gplearn',
       zip_safe=False,
       package_data={'': ['LICENSE']},
       install_requires=['scikit-learn>=0.20.0',
-                        'joblib>=0.13.0'])
+                        'joblib>=0.13.0',
+                        'numpy>=1.17.0'
+                        ])