<!-- Thanks for contributing to gplearn!
Please ensure you have taken a look at the contribution guidelines:
https://gplearn.readthedocs.io/en/stable/contributing.html -->

**Reference Issues/PRs**
<!--
Example: Fixes #1234. See also #3456.
Please use keywords (e.g., Fixes) to create link to the issues or pull requests
you resolved, so that they will automatically be closed when your pull request
is merged. See https://github.com/blog/1506-closing-issues-via-pull-requests
-->

WIP for #188 

**What does this implement/fix? Explain your changes.**

Optimizes speed on old numpy:

numpy 1.16.1:
before changes: 45s
after changes:  38s

numpy 1.17.2:
before changes: 88s
after changes:  87s