@danuker 2019-10-07
Took me a bit more because of SSH keys.
It really does seem numpy 1.17 is much slower, and the fix is negligible :(