@danuker 2020-10-30
Using numpy 1.18.3, my solution is 10 seconds slower. I give up on this.